requirements.txt
annotated-types==0.7.0
anyio==4.11.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
distro==1.9.0
easyocr==1.7.2
fastapi==0.118.0
filelock==3.19.1
filetype==1.2.0
Flask==3.1.2
flask-cors==6.0.1
fsspec==2025.9.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.2
google-auth==2.41.1
google-genai==1.41.0
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.75.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.77
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langsmith==0.4.32
lazy_loader==0.4
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
openai==2.1.0
opencv-python-headless==4.12.0.88
orjson==3.11.3
packaging==25.0
pillow==11.3.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pydantic==2.11.9
pydantic_core==2.33.2
python-bidi==0.6.6
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.16.2
shapely==2.1.2
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.9.30
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
Werkzeug==3.1.3
zstandard==0.25.0
